+++ libc/sysv/consts/ipport.h
@@ -62,7 +62,7 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long IPPORT_USERRESERVED;
+extern const int IPPORT_USERRESERVED;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
